requirements.txt
streamlit
openai
pillow
python-dotenv